+++ packages/extension/src/shared/messages/ShieldMessage.ts
@@ -2,15 +2,14 @@ export type ShieldMessage =
   | { type: "SHIELD_VALIDATE_ACCOUNT" }
   | {
       type: "SHIELD_VALIDATE_ACCOUNT_RES"
-      data: { guardianAddress: string }
     }
   | { type: "SHIELD_VALIDATE_ACCOUNT_REJ"; data: string }
-  | { type: "SHIELD_REQUEST_ACCOUNT_GUARDIAN_ADDRESS" }
+  | { type: "SHIELD_ADD_ACCOUNT" }
   | {
-      type: "SHIELD_REQUEST_ACCOUNT_GUARDIAN_ADDRESS_RES"
+      type: "SHIELD_ADD_ACCOUNT_RES"
       data: { guardianAddress: string }
     }
-  | { type: "SHIELD_REQUEST_ACCOUNT_GUARDIAN_ADDRESS_REJ"; data: string }
+  | { type: "SHIELD_ADD_ACCOUNT_REJ"; data: string }
   | { type: "SHIELD_REQUEST_EMAIL"; data: string }
   | {
       type: "SHIELD_REQUEST_EMAIL_RES"
