+++ packages/extension/src/ui/features/shield/ShieldBaseOTPScreen.tsx
@@ -17,7 +17,12 @@ import {
   emailVerificationStatusErrorSchema,
   getVerificationErrorMessage,
 } from "../../../shared/shield/backend/account"
-import { confirmEmail, requestEmail } from "../../../shared/shield/register"
+import {
+  confirmEmail,
+  requestEmail,
+  shieldAddAccount,
+  shieldValidateAccount,
+} from "../../../shared/shield/register"
 import {
   ShieldValidationErrorMessage,
   getShieldValidationErrorFromBackendError,
@@ -26,7 +31,6 @@ import { updateVerifiedEmail } from "../../../shared/shield/verifiedEmail"
 import { IS_DEV } from "../../../shared/utils/dev"
 import { coerceErrorToString } from "../../../shared/utils/error"
 import { ControlledPinInput } from "../../components/ControlledPinInput"
-import { shieldValidateAccount } from "../../services/shieldAccount"
 import { useYupValidationResolver } from "../settings/useYupValidationResolver"
 import { ShieldValidationErrorScreen } from "./ShieldValidationErrorScreen"
 import { ShieldHeader } from "./ui/ShieldHeader"
@@ -144,8 +148,9 @@ export const ShieldBaseOTPScreen: FC<ShieldBaseOTPScreenProps> = ({
 
                   /** always check the account can be used and exists in backend */
                   await shieldValidateAccount()
+                  await shieldAddAccount()
 
-                  /** successfully verifified with backend - persist this email in the local db */
+                  /** successfully verifified and added account with backend - persist this email in the local db */
                   await updateVerifiedEmail(email)
 
                   onOTPConfirmed()
